+++ rx-location-manager-nmea/src/main/kotlin/net/samystudio/rxlocationmanager/nmea/GGA.kt
@@ -1,5 +1,7 @@
 package net.samystudio.rxlocationmanager.nmea
 
+import java.util.*
+
 class GGA(message: String) : Nmea(message) {
     val time: String by lazy { data[1] }
     val latitude: Double? by lazy {
@@ -28,6 +30,40 @@ class GGA(message: String) : Nmea(message) {
     val ellipsoidalOffset: Double? by lazy { data[11].toDoubleOrNull() }
     val differentialGpsAge: Double? by lazy { data[13].toDoubleOrNull() }
     val differentialGpsStationId: String by lazy { data[14] }
+
+    constructor(
+        type: String,
+        time: String = "",
+        latitude: Double? = null,
+        longitude: Double? = null,
+        quality: Quality = Quality.NO_FIX,
+        satelliteCount: Int = 0,
+        horizontalDilutionOfPrecision: Double? = null,
+        altitude: Double? = null,
+        ellipsoidalOffset: Double? = null,
+        differentialGpsAge: Double? = null,
+        differentialGpsStationId: String = ""
+    ) : this(
+        "$%sGGA,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s".format(
+            type.toUpperCase(Locale.ROOT),
+            time,
+            latitude?.let { convertLocationNmea(it) } ?: "",
+            latitude?.let { if (it < 0) LocationDirection.S.name else LocationDirection.N.name }
+                ?: "",
+            longitude?.let { convertLocationNmea(it) } ?: "",
+            longitude?.let { if (it < 0) LocationDirection.W.name else LocationDirection.E.name }
+                ?: "",
+            Quality.values().indexOf(quality),
+            satelliteCount,
+            horizontalDilutionOfPrecision ?: "",
+            altitude ?: "",
+            altitude?.let { 'M' } ?: "",
+            ellipsoidalOffset ?: "",
+            ellipsoidalOffset?.let { 'M' } ?: "",
+            differentialGpsAge ?: "",
+            differentialGpsStationId
+        ).let { "%s*%s".format(it, computeChecksum(it)) }
+    )
 
     override fun getTokenValidators(): Array<TokenValidator> {
         val optionalDoubleValidator = DoubleValidator(true)