+++ rx-location-manager-nmea/src/test/kotlin/net/samystudio/rxlocationmanager/nmea/RMCNavigationalTest.kt
@@ -30,6 +30,29 @@ class RMCNavigationalTest {
     }
 
     @Test
+    fun validateAlternateConstructor() {
+        val rmc = RMCNavigational(
+            "GN",
+            "191613.43",
+            RMCTiming.Status.A,
+            33.71103,
+            -117.85643,
+            5.0,
+            10.2,
+            "100520",
+            9.3,
+            LocationDirection.W,
+            RMCTiming.Mode.A,
+            RMCNavigational.NavigationalStatus.S
+        )
+
+        assertEquals(
+            "\$GNRMC,191613.43,A,3342.6618,N,11751.3858,W,5.0,10.2,100520,9.3,W,A,S*6E",
+            rmc.message
+        )
+    }
+
+    @Test
     fun validateEmpty() {
         val rmc = RMCNavigational("\$GPRMC,,V,,,,,,,,,,N,V*29")